+++ src/main/java/br/com/fiap/global/baze/repository/DesafioRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.fiap.global.baze.model.Desafio;
 
-public interface DesafioRepository extends JpaRepository<Desafio, Integer> {
+public interface DesafioRepository extends JpaRepository<Desafio, Long> {
     
 }
