+++ src/main/java/br/com/fiap/global/baze/repository/TelefoneRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.fiap.global.baze.model.Telefone;
 
-public interface TelefoneRepository extends JpaRepository<Telefone, Integer> {
+public interface TelefoneRepository extends JpaRepository<Telefone, Long> {
     
 }
